feat : 디렉토리 부모의 이름까지 물고 response로 보내기 구현 완료

+++ backend/src/main/java/com/ssafy/getsbee/domain/directory/service/DirectoryServiceImpl.java
@@ -23,6 +23,8 @@ public class DirectoryServiceImpl implements DirectoryService {
     private final DirectoryRepository directoryRepository;
     private final MemberRepository memberRepository;
 
+    private int ROOT_DEPTH = 0;
+
     @Override
     public List<DirectoryResponse> findAllByMember(Member member) {
         List<Directory> directories = directoryRepository.findAllByMember(member);
@@ -59,9 +61,11 @@ public class DirectoryServiceImpl implements DirectoryService {
 
 
         for(Directory directory : existingDirectories){
+            if(directory.getDepth()==ROOT_DEPTH) continue;
             if(!requestDirectoryIds.contains(directory.getId())){
                 if(directory.getName().equals("Bookmark") || directory.getName().equals("Temporary")||
                         directory.getName().equals("Root")) {
+                    System.out.println("deleting: " + directory.getName());
                     throw new BadRequestException(CANT_DELETE_DEFAULT_DIRECTORY);
                 }
                 directoryRepository.delete(directory);
@@ -102,8 +106,7 @@ public class DirectoryServiceImpl implements DirectoryService {
     @Override
     public String findFullNameByDirectory(Directory directory) {
         if(directory.getDepth()==1) return directory.getName();
-        return directory.getParentDirectory().getName() + " / " +directory.getName()
-        return "";
+        return directory.getParentDirectory().getName() + " / " +directory.getName();
     }
 
     private void filterDirectoriesByAuth(List<Directory> directories) {
